package/python3: drop unnecessary patches
Patches 0004 and 0009 relied upon environment variables that were used
to drive custom logic in distutils when executed by an external Python
interpreter. This issue should have been mostly addressed upstream [0],
but even if that was not the case, distutils is no longer a supported
installation method for packages to use for installation as of 09de823c.
As such, there shouldn't be a need to continue to patch it as setuptools
vendors its own distutils that supercedes the one provided by Python.

Patches 0010 and 0026 (see [1] [2] [3]) can be removed due to a fix
introduced in 3.7 [4].

Patch 0028 can be dropped since commit 3fed4245 removed the dependency
on a host provided python to build host-python3.

[0] https://github.com/python/cpython/commit/9731330d6f4d63d6d57c8e99c8d11ef1def42858
[1]: https://github.com/python/cpython/issues/67096
[2]: https://github.com/python/cpython/issues/75894
[3]: https://github.com/python/cpython/issues/66338
[4]: https://github.com/python/cpython/commit/14086cfc5eed8c5e78342d79e5db87a135d75fa8

Signed-off-by: Vincent Fazio <vfazio@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>