package/ivi-homescreen: drop bogus comment dependency on NPTL
As part of the review process of the ivi-homescreen package, which
landed in commit 9d8497e79de3ced144877ba6ee38cc5b7181aaeb, it was
deemed that the dependency on BR2_TOOLCHBAIN_HAS_THREAD_NPTL was not
needed, as it was implied by glibc. According to the commit log:

    [yann.morin.1998@free.fr:
      - propagate BR2_PACKAGE_HOST_FLUTTER_SDK_BIN_ARCH_SUPPORTS to comments
      - drop NPTL, implied by glibc
      - reorder dependencies in a more logical way
      - reorder comments
      - drop undefined BR2_PACKAGE_IVI_HOMESCREEN_HAS_CLIENT
      - grammar ("for to change")
    ]
    Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>

However, while the BR2_TOOLCHBAIN_HAS_THREAD_NPTL dependency was
removed from the BR2_PACKAGE_IVI_HOMESCREEN option definition, the
corresponding dependency in the Config.in comment was not
dropped. Let's bring things back in sync.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>