package/libnl: update the patch to be applied with fuzz 0
This commit allows the package patch to be applied with fuzz factor 0.
The fuzz factor specifies how many lines of the patch can be inexactly
matched, so the value 0 requires all lines to be exactly matched.

While we're at it, also fix the upstream status to the proper Upstream:
tag.

Signed-off-by: Dario Binacchi <dario.binacchi@amarulasolutions.com>
[Arnout: add proper Upstream: tag and update checkpackageignore]
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>