package/mosquitto: build mosquitto_{ctrl,passwd} if the broker is enabled
The tools require openssl (and cjson for mosquitto_ctrl), and are silently
skipped by the mosquitto build system if those are not available.

The tools are small compared to the broker and dependencies:

-rwxr-xr-x 1 peko peko 4.8M Jun  1 00:42 target/usr/lib/libcrypto.so.3
-rwxr-xr-x 1 peko peko 968K Jun  1 00:42 target/usr/lib/libssl.so.3
-rwxr-xr-x 1 peko peko 266K Jun  1 00:42 target/usr/sbin/mosquitto
-rwxr-xr-x 1 peko peko  54K Jun  1 00:42 target/usr/bin/mosquitto_ctrl
-rwxr-xr-x 1 peko peko  34K Jun  1 00:42 target/usr/lib/libcjson.so.1.7.18
-rwxr-xr-x 1 peko peko  26K Jun  1 00:42 target/usr/bin/mosquitto_passwd

So automatically build them (if the dependencies are available) when the
broker is built.

Notice: We have to explicitly build apps/mosquitto_ctrl +
apps/mosquitto_passwd and not just apps, as apps/Makefile uses a DIRS
variable that conflicts with the DIRS=..  we pass to the toplevel Makefile.

Signed-off-by: Sebastien Laveze <slaveze@smartandconnective.com>
[Peter: drop dedicated option, also handle mosquitto_ctrl]
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>