package/libgit2: bump version to 1.8.1
Upstream notes that libhttp-parser is unmaintained and suggest llhttp
as a replacement.  This version bundles llhttp (and includes its license
in COPYING) but still allows to use a system-provided libhttp-parser.

llhttp (https://github.com/nodejs/llhttp) is a C library that depends on
a nodejs interpreter and more than 10 npm dependencies to generate its
parser code.  The bundled version contains the generated sources and
could be built without nodejs.  But distributing this generated code
as part of libgit2 would probably be legally dubious because of the
GPL's original source code provision, even if llhttp is MIT-licensed.

Packaging llhttp would be a daunting task and its build will require
considerable time and space, so for now, keep using libhttp-parser.

Signed-off-by: Nicolas Cavallari <nicolas.cavallari@green-communications.fr>
Signed-off-by: Arnout Vandecappelle <arnout@mind.be>