configs/zynq_*: move hashes
This patch moves the linux and uboot hashes from 'board/zynq/patches' to
'board/xilinx/patches'.  The reason for this is that all of the Xilinx boards
including zynq, zynqmp and versal will be using the same linux and uboot repo.
By putting the hashes in 'board/xilinx/patches', the same hashes can be shared
amongst all of the Xilinx boards.

Signed-off-by: Neal Frager <neal.frager@amd.com>
Reviewed-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>